Add Fedbird project
Fedbird project is an object detection project on Seabird dataset
donated to AI Sweden. It uses Tiny Yolo3 model for object detection,
and it is a showcase for HelloChina usecase in the FedML testbed project.

Added seed model and the package as well in case someone wants to quickly
try it but it should not be needed.